Adding in a contract for df-get-series
Although it is not intended to be exposed, I need it if I'm going to extend things in the dataframe.

Ultimately, I'll discuss whether there are better ways, or if this is acceptable as a pull requst.

For the moment, this was the minimal change I could make that would allow me to continue exploring
some of my own ideas with the data-frame as something underneath my code.